Refactorizacion del codigo utilizando el patron Container/Presentational
