+++ OrderNow-React/.env.example
@@ -1,3 +1,3 @@
-VITE_SUPABASE_KEY=TU-KEY-SUPABASE
-VITE_SUPABASE_URL=TU-URL-SUPABASE
+VITE_SUPABASE_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImluYmNsZm1idGp6aml1bHp1ZG12Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDM0NDkwMTMsImV4cCI6MjA1OTAyNTAxM30.cgw-RaBDW3X9DnsPBKZikSEN7ztakfSDmat9ilhMU9o
+VITE_SUPABASE_URL=https://inbclfmbtjzjiulzudmv.supabase.co
 VITE_BACKEND_ST=wkjasdkljflfksj.com
